Move :extensions into :sdk:core:extensions
