--- components/extensions/api/extensions.api
@@ -1,7 +1,0 @@
-public final class io/github/composegears/valkyrie/extensions/PathUtilsKt {
-	public static final fun writeToKt (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ZZ)V
-	public static synthetic fun writeToKt$default (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ZZILjava/lang/Object;)V
-	public static final fun writeToXml (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ZZ)V
-	public static synthetic fun writeToXml$default (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ZZILjava/lang/Object;)V
-}
-
