--- components/extensions/api/extensions.klib.api
@@ -1,10 +1,0 @@
-// Klib ABI Dump
-// Targets: [wasmJs]
-// Rendering settings:
-// - Signature version: 2
-// - Show manifest properties: true
-// - Show declarations: true
-
-// Library unique name: <valkyrie.components:extensions>
-final inline fun <#A: reified kotlin/Any> (kotlin/Any?).io.github.composegears.valkyrie.extensions/cast(): #A // io.github.composegears.valkyrie.extensions/cast|cast@kotlin.Any?(){0§<kotlin.Any>}[0]
-final inline fun <#A: reified kotlin/Any> (kotlin/Any?).io.github.composegears.valkyrie.extensions/safeAs(): #A? // io.github.composegears.valkyrie.extensions/safeAs|safeAs@kotlin.Any?(){0§<kotlin.Any>}[0]
