+++ components/psi/imagevector/src/main/kotlin/io/github/composegears/valkyrie/psi/imagevector/common/KtCallExpression.kt
@@ -1,12 +1,12 @@
 package io.github.composegears.valkyrie.psi.imagevector.common
 
-import io.github.composegears.valkyrie.extensions.safeAs
 import io.github.composegears.valkyrie.ir.IrColor
 import io.github.composegears.valkyrie.ir.IrFill
 import io.github.composegears.valkyrie.ir.IrPathFillType
 import io.github.composegears.valkyrie.ir.IrStroke
 import io.github.composegears.valkyrie.ir.IrStrokeLineCap
 import io.github.composegears.valkyrie.ir.IrStrokeLineJoin
+import io.github.composegears.valkyrie.sdk.core.extensions.safeAs
 import org.jetbrains.kotlin.psi.KtCallExpression
 
 internal fun KtCallExpression.extractPathFillType(): IrPathFillType {
