+++ components/psi/imagevector/src/test/kotlin/io/github/composegears/valkyrie/psi/imagevector/common/KtFile.kt
@@ -3,8 +3,8 @@ package io.github.composegears.valkyrie.psi.imagevector.common
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiManager
 import com.intellij.testFramework.LightVirtualFile
-import io.github.composegears.valkyrie.extensions.cast
 import io.github.composegears.valkyrie.resource.loader.ResourceLoader.getResourceText
+import io.github.composegears.valkyrie.sdk.core.extensions.cast
 import org.jetbrains.kotlin.idea.KotlinFileType
 import org.jetbrains.kotlin.psi.KtFile
