+++ sdk/core/extensions/api/extensions.api
@@ -1,0 +1,5 @@
+public final class io/github/composegears/valkyrie/sdk/core/extensions/PathUtilsKt {
+	public static final fun writeToKt (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ZZ)V
+	public static synthetic fun writeToKt$default (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;ZZILjava/lang/Object;)V
+}
+
