+++ sdk/core/extensions/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/core/extensions/Any.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.extensions
+package io.github.composegears.valkyrie.sdk.core.extensions
 
 inline fun <reified T : Any> Any?.safeAs(): T? = this as? T
