+++ tools/compose-app/src/wasmJsMain/kotlin/io/github/composegears/valkyrie/main.kt
@@ -5,7 +5,7 @@ package io.github.composegears.valkyrie
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
-import io.github.composegears.valkyrie.extensions.safeAs
+import io.github.composegears.valkyrie.sdk.core.extensions.safeAs
 import kotlinx.browser.document
 import org.w3c.dom.HTMLElement
