+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/editor/VirtualFile.kt
@@ -3,7 +3,7 @@ package io.github.composegears.valkyrie.editor
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiFileFactory
-import io.github.composegears.valkyrie.extensions.safeAs
+import io.github.composegears.valkyrie.sdk.core.extensions.safeAs
 import org.jetbrains.kotlin.idea.KotlinFileType
 import org.jetbrains.kotlin.psi.KtFile
