+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/di/DI.kt
@@ -2,7 +2,7 @@ package io.github.composegears.valkyrie.ui.di
 
 import com.composegears.leviathan.ProvidableDependency
 import com.intellij.openapi.project.Project
-import io.github.composegears.valkyrie.extensions.cast
+import io.github.composegears.valkyrie.sdk.core.extensions.cast
 
 object DI {
     private val platformModule = intellijPlatformModule()
