+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/platform/DragAndDropHandler.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.LocalInspectionMode
 import io.github.composegears.valkyrie.compose.core.rememberMutableState
-import io.github.composegears.valkyrie.extensions.safeAs
+import io.github.composegears.valkyrie.sdk.core.extensions.safeAs
 import io.github.composegears.valkyrie.ui.foundation.theme.LocalComponent
 import io.github.composegears.valkyrie.ui.platform.DragAndDropHandlerState.Companion.dragging
 import io.github.composegears.valkyrie.ui.platform.DragAndDropHandlerState.Companion.notDragging
