--- tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/util/TiamatSavedState.kt
@@ -1,6 +1,0 @@
-package io.github.composegears.valkyrie.util
-
-import com.composegears.tiamat.navigation.SavedState
-import io.github.composegears.valkyrie.extensions.safeAs
-
-inline fun <reified T : Any> SavedState.getOrNull(key: String): T? = get(key).safeAs<T>()
